[Fix] Merge from Account into develop

See merge request s03-webmobile2-sub2/s03p12a106!167

+++ frontend/src/components/account/mine/EditMyPageHeader.vue
@@ -9,8 +9,7 @@
       <p class="text-dark mb-0"><strong>프로필 편집</strong></p>
     </b-nav-item>
     <b-nav-item class="d-flex justify-content-end">
-      <button @click="editprofile" >
-        <!-- <button @click="editprofile" v-show="isValid"> -->
+        <button @click="editprofile" v-show="is_valid.is_valid=='is-valid'">
       <b-icon-check2-circle class="text-dark"></b-icon-check2-circle>
       </button>
     </b-nav-item>
@@ -22,40 +21,49 @@ import AccountApi from '@/api/AccountApi';
 
 export default {
   name: 'EditMyPageHeader',
-  data() {
-    return {
-      isValid : this.valid,
-    }
-  },
   props: {
     info: Object,
-    valid: Boolean
-  },
-  mounted() {
-    console.log(this.valid);
+    is_valid: Object
   },
   methods: {
     backpage(){
-      if(confirm("저장되지 않은 변경 내용이 있습니다. \n 취소하시겠어요?"))
+      if(JSON.stringify(this.info) != JSON.stringify(this.$store.state.userInfo)){
+        if(confirm("저장되지 않은 변경 내용이 있습니다. \n 취소하시겠어요?"))
+          this.$router.go(-1);
+      }else{
         this.$router.go(-1);
+      }
     },
     editprofile(){
-      AccountApi.requestModifyProfile(
-        this.info,
-        res => {
-          if(res.success){
-            this.$store.commit('addUserInfo', res.user);
-            localStorage.setItem('tvility', JSON.stringify(res.user));
-            alert('변경되었습니다');
-          }else{
-            alert('변경에 실패했습니다');
+      if(JSON.stringify(this.info) == JSON.stringify(this.$store.state.userInfo)){
+        this.makeToast("변경사항이 없습니다.", "warning");
+      }else if(this.valid){
+        AccountApi.requestModifyProfile(
+          this.info,
+          res => {
+            if(res.success){
+              this.$store.commit('addUserInfo', res.user);
+              localStorage.setItem('tvility', JSON.stringify(res.user));
+              this.makeToast("변경되었습니다.", "primary");
+            }else{
+              this.makeToast("변경에 실패했습니다.", "danger")
+            }
+          },
+          error => {
+            this.$router.push({name:'Errors', query: {message: error.msg}})
           }
-        },
-        error => {
-          this.$router.push({name:'Errors', query: {message: error.msg}})
-        }
-      );
-    }
+        );
+      }
+    },
+    makeToast(message, variant){
+        this.$bvToast.toast(message, {
+          title: '알림',
+          toaster: "b-toaster-bottom-right",
+          variant: variant,
+          autoHideDelay: 3000,
+          appendToast: false
+        })
+     }
   },
 }
 </script>
